requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
elevenlabs-sdk
ffmpeg-python